Fix some utils functions
